modal: add large mode
set attribute large to true when creating a modal.
It's 740px wide (default width is 660px). 740px because the minimum supported width is 1024px (1280px recommended). The sidebar is 260px wide, leaving 764px for a local modal. 740px results in a minimum padding of 12px around the modal window inside the content window.