Use deliver_method :test for tests
This avoids flooding the screen with warnings on systems without MTA:

Please install an MTA on this system if you want to use sendmail!
Please install an MTA on this system if you want to use sendmail!
Please install an MTA on this system if you want to use sendmail!