Better sidebar scrollbar spacing
With the sidebar tabs overlapping the main content there needs to be a gap between the sidebar and the tabs for the scrollbar. The most elegant solution would be to only open up this gap when a scrollbar is visible. But the tabs are part of the sidebar and thus life in a different controller. For now a cleaner solution that to force a scrollbar is to show an empty scrollbar.