allow agents to always reopen tickets, regardless of group.follow_up_possible
This commit changes the behavior of the option group.follow_up_possible
(admin -> manage -> groups -> <group> -> follow up possible). Before this
commit, setting this option to "do not reopen ticket but create a new ticket"
meant that _both_ agents and customers weren't allowed to reopen closed
tickets. This is problematic because it means that an agent who accidentally
closed a ticket has no way to undo this, which I believe wasn't the intention
of this option.

This commit changes the behavior so that agents are now exempt from this
limitation and can close and reopen tickets at whim.