Make sure that restoration script sets correct rights on files
These changes ensure that, no matter if migrating from a online instance or importing some other Zammad-Export, that the filesystem rights on the Zammad-Directory are always correct.

We're using fixed user "zammad" here, the Backup-Documentation needs enhancement and will hold a warning for this to ensure, other users don't crash here.