Add RocketChat integration
This commit adds RocketChat integration, mimics Slack.

It provides browser configuration to request notifications on
ticket create/update/etc. Notifications are sent on defined
channel. Configured user must be access to this channel.

Additionally, if article contains a user mention with @userid
(one or many), a direct message is sent to specified users help
with message templates. By defaults it sends information that user
is mentioned in ticket XXX with text of note.