Fix deserializtion of multiselect value
When a multiselect Attribute is used with the Ticket object, the rails
backend persists the multiselect value as a serialized JSON array in the
database.

When this value is read back and passed down to the form widget to
render, it -- before this patch -- was simply rendering the serialized
JSON array string as an option.

This patch fixes the widget so that it will attempt to deserialize the
value as an array to render the multi selected options properly.

For the case where the passed value serializes to something other than
an array, it falls back on the default existing behavior of just
rendering the string as is.

In the event some bad data gets persisted to the database, we do not
want to break the UI. So we swallow JSON parse errors and fall back on
the existing default behavior of just rendering the string as is.