Testing: Spec Channel::Driver::Twitter#send
This commit is part of a larger effort to overhaul test coverage
for Twitter-related logic throughout the application.

The logic tested in this commit _could_ have been tested
at other interfaces:

* Channel#deliver
  (higher abstraction; calls Channel::Driver::Twitter#send)

* TwitterSync#from_article
  (lower abstraction; is called from Channel::Driver::Twitter#send)

So why did we choose to test at the channel driver level?

Channel#deliver accepts many different adapters
(of which Channel::Driver::Twitter is just one).
Testing Channel#deliver with every possible combination of inputs
would quickly lead to a single, unmanageably large test file
with an explosion of context blocks.
Testing each adapter in isolation seemed the wiser approach.

TwitterSync is a helper class without a cohesive single responsibility,
and thus is strong candidate for refactoring.
It is safer to change its interface than the channel driver's.