Testing: Fix false positive and cover edge case for Channel::Driver::Twitter#fetch
Originally, a spec for race condition handling
in Channel::Driver::Twitter#fetch
used a `travel_back` call in the wrong place:
it needed to be called _after_ `subject!(:channel)` was created,
but since it was placed in an `around` block,
it ran before all preceding setup blocks.

Moving it to a `before` block changed the order in which it was called,
thus fixing a false positive in the test.

An additional expectation was added
to prevent this false positive in the future,
and a new test case was added to capture yet another edge case.