Fixes #3160 - Backup script restoration order causes issues if version difference causes ruby version differences
This commit ensures that the restore script will restore the database **before** restoring the file system of Zammad.

This solves ruby errors for environments that don't have old ruby versions installed (because they shouldn't be required during restorations).

This commit provides a compromise and behaves the old way for docker and source code installations. This is because it may be a source code installation is restored right away and was not installed before.

Hints for this are prepared for the documentation over at: https://github.com/zammad/zammad-documentation/pull/140