Fixes #3139 - Restore script uses postgres user and nukes restoration of hosted dumps
Previously restoring Zammad dumps (like those from a Hosted setup) couldn't be imported successfully, because the "owner" information is not provided by the dump. If that's the case, postgre SQL will use the current user you're importing with. If that's postgres and not Zammad, you can't read and use the database as Zammad user.