Fixes #3137 - New email account expert view cannot be opened without filling in all fields
This fix tries to balance:

1. Improving the user experience when a user wants to input
   expert/advanced settings directly.
2. Do as little code changes as possible.
3. Leave the internal state as is.

A such, the fix assumes that the 'Experts' form is an extension of the
base one instead of being additional settings and shows all the relevant
settings from the base form in the expert view too. This allows the user
to directly open the experts view when adding a new email. Already entered
values are not lost and related fields are filled in.

Pros of this approach:

- In the basic and expert case, the user has only one form to fill in
  with all the relevant settings.
- The validation of the form does not need special handling.
- The user gets immediate feedback about validation errors.
- Housekeeping on the code side is minimal.

Cons

- An additional array, which is semantically but not syntactically linked
  to the `configureAttributesBase` array is created (email.coffee:416-418)
- `@account.meta` needs manual update to be in sync (email.coffee:553-556)

The last con could be not needed but because of (3) and my unfamiliarity
with the implications of having `@account.meta` out-of-sync, I opt-in
to update it.

Another approach considered was moving the validation of both forms to
occur when the 'submit' button of the expert form was pressed. This would
get rid of the mentioned cons but introduce jarring behaviour for the user
when a base form validation failed on the expert view and the modal would
change back.

Potential issues:

- Going back from the expert view does not carry over related settings
  as it does going from the base view to the expert view. This is in-line
  with the behaviour before this fix.
- Going to the expert view does not use its own callback but reuses
  `probeBasedOnIntro` (email.coffee:484)
- Changing the 'SSL/STARTTLS' setting to 'yes' in the expert view does
  not set the port if the port is 143 (the default for no SSL).
  Changing the setting to 'no' does set the port.

The last point is technically not in scope of the issue and the other
points would need more code changes than strictly necessary for
functionality, violating (2). However, those small further
improvements could result in a net win and could be added with further
commits.