Partial-fix #3460 - 'Drop Files Here' not always hiding
Selecting, dragging & dropping a text into the text-field under
`/#ticket/zoom/:id` calls `onDrop` from html5Upload.js:104.
This sets `inCounter` to 0 but call calls `onDragEnter()` afterwards,
followed by the method to process files. (As only text was dragged, no
files will be processed.)
In the end, this sequence of events result in `inCounter` being 1 instead
of 0 as `onDragEnter()` increments the counter by one. I imagine the
invariant is that each `onDragEnter()` is followed by a `onDragLeave()`,
which is broken in this case.

As `inCounter` starts at 1 instead of 0, the next time a text is dragged
around and dropped somewhere, the 'Drop Files Here' area will not be
hidden. This commit fixes that by moving `inCounter = 0` in the
`onDrop(e)` method after `onDragEnter(e)`, instead of before.
This is only a partial fix for #3460, as the text-field under
`/#ticket/create/id/:id` seems still to be afflicted which seems to
indicate more issues.

I am also not sure how html5Upload.js and jquery.contenteditable.js work
together, the latter being the source of a null-exception when the
'Drop Files Here' issue occur on `/#ticket/create/id/:id`.
There seems to be a deeper conflict with 'false' drops of selected text
instead of upload-able content.