Fixes #3460 - "Drop files here" drag area not always hiding
When dragging & dropping e.g. selected text on the text-field under
`#ticket/create/id/:id` the 'drop' event is usually captured by the
'dropArea' and bubbled up until the to the form attached event handler
captures it and calls `onDrop(e)` of html5Upload.js.
However, the area of the div with the id 'TicketArticle_:id_body' is
slightly larger as the area of the 'dropArea' div.
Those two divs are siblings to one another and the
'TicketArticle_:id_body' has a 'drop' event listener to
`Plugin.onDrop(e)` from jquery.contenteditable.js:414, which captures
the event and stops propagation, if the user let go of a drag just
directly below the 'dropArea' (a few pixels there belong to
'TicketArticle_:id_body').
This (a) throws a null-exception as there is no file to upload if text
was dragged and (b) stops propagation and in consequence, the removal
of the 'dropArea', blocking input.

This commit fixes it by explicitly checking for null in `Plugin.onDrop(e)`,
however, changing the sizes of the divs could also be a viable fix.

It may be needed to re-evaluate the whole drag'n'drop mechanism but my
unfamiliarity with the code base leaves me hesitant to do any big changes.