LDAP: apply user filter during auth on bind
Previously the user filter was not used for the login attempt.
Thus in case of multiple LDAP entries matching the login name (e.g. `cn=foo`), the wrong entry could be used (by default the first result is attempted).

In our case we have an LDAP tree with an "identity" and an "account" for each user (both residing in different branches of the LDAP tree).

The `user_filter` needs to be applied during login attempts in order to avoid this ambiguity.