Shim STARTTLS support
Added very simple STARTTLS support based on protocol in the host_url field. The Ruby LDAP module already supports start_tls so the only thing required was the logic to select it as the encryption method.

Port is also automatically determined based on common defaults for each protocol, or via the normal @port value if given in the config explicitly.