Replace 'gzip' with 'zstd' compression algorithm
An idea originally conceived by @sigio and put to use in our local
Zammad instances. We initially replaced `gzip` with `pigz`, to add
multi-threading to the compressor, alleviating the workload of a single
CPU and spreading it out over all of our cores instead. A feature which
we *severely* missed in some of our *large* Zammad setups. We decided to
further incorporate the Zstandard compression algorithm and to
completely replace `gzip` with `zstd`, bringing the compression
algorithm into the 21st century.

This feature should be considered as a (somewhat) *breaking change*, as
the new backup script won't be able to restore the older gzipped files
anymore with these changes in place. Either re-compress older backups,
keep an archived copy of the `functions` script to be able to (more
easily) restore older backups, or restore manually. It's not that
difficult.

You'll probably want to clean up the old gzipped fileformat archives in
your backup directory as well, as the new script only touches the new
`.zst` files.

Package dependencies haven't been updated, as this is an optional
script. `zstd` seems to be installed by default(?) on current versions
of Debian and Ubuntu. It should be installed manually on distros such as
RHEL/CentOS/etc.